Move options under 'Edit Profile' to LoggedInNavbar, and get rid of EditProfile
