+++ src/components/GlobalNavbar/LoggedInNav.tsx
@@ -10,6 +10,8 @@ import {
   useAuthenticatedUser,
   useAuthenticatedUserAddress,
 } from 'hooks/api/users/useUser';
+import { useModal } from 'contexts/modal/ModalContext';
+import EditUserInfoModal from 'scenes/UserGalleryPage/EditUserInfoModal';
 
 function truncate(address: string) {
   return `${address.slice(0, 4)}...${address.slice(-4)}`;
@@ -19,6 +21,7 @@ function LoggedInNav() {
   const { logOut } = useAuthActions();
   const user = useAuthenticatedUser();
   const userAddress = useAuthenticatedUserAddress();
+  const { showModal } = useModal();
 
   const truncatedUserAddress = useMemo(() => {
     return truncate(userAddress);
@@ -28,11 +31,31 @@ function LoggedInNav() {
     navigate(`/${user.username}`);
   }, [user.username]);
 
+  const handleEditNameClick = useCallback(() => {
+    showModal(<EditUserInfoModal />);
+  }, [showModal]);
+
+  const handleEditGalleryClick = useCallback(() => {
+    navigate('/edit');
+  }, []);
+
   return (
     <Container>
       <TextButton onClick={handleGalleryRedirect} text="My Gallery" />
       <Spacer width={24} />
       <Dropdown mainText="Account">
+        <TextButton
+          text="Edit Gallery"
+          onClick={handleEditGalleryClick}
+          underlineOnHover
+        />
+        <Spacer height={12} />
+        <TextButton
+          text="Edit name & Bio"
+          onClick={handleEditNameClick}
+          underlineOnHover
+        />
+        <Spacer height={12} />
         <CopyToClipboard textToCopy={userAddress}>
           <TextButton
             text={truncatedUserAddress}
