+++ src/scenes/UserGalleryPage/UserGalleryHeader.tsx
@@ -1,13 +1,7 @@
-import { useCallback } from 'react';
-import { navigate } from '@reach/router';
 import styled from 'styled-components';
 import { Subdisplay, BodyRegular } from 'components/core/Text/Text';
 import Spacer from 'components/core/Spacer/Spacer';
 import colors from 'components/core/colors';
-import Dropdown from 'components/core/Dropdown/Dropdown';
-import TextButton from 'components/core/Button/TextButton';
-import { useModal } from 'contexts/modal/ModalContext';
-import EditUserInfoModal from './EditUserInfoModal';
 import { User } from 'types/User';
 
 type Props = {
@@ -16,40 +10,11 @@ type Props = {
 };
 
 function UserGalleryHeader({ user, isAuthenticatedUsersPage }: Props) {
-  const { showModal } = useModal();
-
-  const handleEditNameClick = useCallback(() => {
-    showModal(<EditUserInfoModal />);
-  }, [showModal]);
-
-  const handleEditGalleryClick = useCallback(() => {
-    navigate('/edit');
-  }, []);
-
   return (
     <StyledUserGalleryHeader>
       <Subdisplay>{user.username}</Subdisplay>
       <Spacer height={12} />
       <StyledUserDetails>
-        {isAuthenticatedUsersPage && (
-          <StyledRow>
-            <StyledRightContainer>
-              <Dropdown mainText="Edit Profile">
-                <TextButton
-                  text="Edit Gallery"
-                  onClick={handleEditGalleryClick}
-                  underlineOnHover
-                />
-                <Spacer height={12} />
-                <TextButton
-                  text="Edit name & Bio"
-                  onClick={handleEditNameClick}
-                  underlineOnHover
-                />
-              </Dropdown>
-            </StyledRightContainer>
-          </StyledRow>
-        )}
         <Spacer height={18} />
         <StyledLeftContainer>
           {/* TODO: won't be able to determine `MemberSince` unless we crawl blockchain 
@@ -63,11 +28,6 @@ function UserGalleryHeader({ user, isAuthenticatedUsersPage }: Props) {
     </StyledUserGalleryHeader>
   );
 }
-
-const StyledRow = styled.div`
-  display: flex;
-  justify-content: flex-end;
-`;
 
 const StyledUserGalleryHeader = styled.div`
   display: flex;
@@ -87,10 +47,6 @@ const StyledLeftContainer = styled.div``;
 const StyledBodyRegular = styled(BodyRegular)`
   /* ensures linebreaks are reflected in UI */
   white-space: pre-wrap;
-`;
-
-const StyledRightContainer = styled.div`
-  display: flex;
 `;
 
 export default UserGalleryHeader;
